CORS/django_app/src/music/viewsets.py
from rest_framework import viewsets

from music.models import Album
from music.serializers import AlbumSerializer


class AlbumViewSet(viewsets.ModelViewSet):
    queryset = Album.objects.all()
    serializer_class = AlbumSerializer
