CORS/django_app/requirements.txt
django==2.1.5
django-cors-headers==2.5.2
djangorestframework
